// src/shared/api/data-contracts.ts
export interface UserResponse {
  /** User id */
  id: number;
  /** First name */
  first_name: string;
  /** Second name */
  second_name: string;
  /** Display name */
  display_name: string;
  /** User login - unique */
  login: string;
  /**
   * User email - unique
   * @format email
   */
  email: string;
  /** User phone */
  phone: string;
  /** Avatar */
  avatar: string;
}
// ...
export interface HttpErrorBody {
  /** Error message */
  reason: string;
}

export interface UserUpdateRequest {
  /** First name */
  first_name?: string;
  /** Second name */
  second_name?: string;
  /** Display Name */
  display_name?: string;
  /** User login - unique */
  login?: string;
  /** Email */
  email?: string;
  /** Phone */
  phone?: string;
}

export interface UserRequest {
  /** First name */
  first_name: string;
  /** Second name */
  second_name: string;
  /** Display Name */
  display_name: string;
  /** User login - unique */
  login: string;
  /** Email */
  email: string;
  /** Phone */
  phone: string;
}

export interface FindUserRequest {
  /** User login (beginning of login) */
  login: string;
}

export interface ChangePasswordRequest {
  /** Old password */
  oldPassword: string;
  /** New password */
  newPassword: string;
}